Add always relative import style setting to flake8-tidy-imports
This commit changes the flake8-tidy-imports TID252 rule and its
ban-relative-imports setting to allow forcing imports to be always
absolute, relative, or absolute for parents.

Besides adding the always-relative mode to this setting, this commit: -
Rename `ban_relative_imports` to `relative_import_style` in `Settings`
and `Options`. - Rename Strictness to ImportStyle - Update the
diagnostic messages and fixes to align with the `ImportStyle` enum. -
Update the code to use `ImportStyle` in
`flake8_tidy_imports::rules::relative_imports`. - Update the code in
`ruff_linter/src/checkers/ast/analyze/statement.rs` to use
`ImportStyle`.

The commit relies on the categorize function used for the isort rule to
determine if absolute imports are first party and therefore could be
relative.

IMPORTANT: I realize now that this implementation is wrong. Users can
configure separate packages to be 'first party', but they can't be
imported relatively. I'll have to take a closer look at the ImportType
options and see which one applies to same-package imports that could
become `from .bar import baz` instead of `from foo.bar import baz`.