`Option::unwrap` is now const
Summary
--

I noticed while working on #20006 that we had a custom `unwrap` function for
`Option`. This has been const on stable since 1.83 (docs, release notes), so I
think it's safe to use now. I grepped a bit for related todos and found this one
for `AsciiCharSet` but no others.

Test Plan
--

Existing tests